+++ apps/analysis-api/src/errors/body-required.error.ts
@@ -1,0 +1,9 @@
+import { AbstractError } from './abstract.error';
+
+export class BodyRequiredError extends AbstractError {
+
+    constructor() {
+        super('Missing message in your request body', 400);
+        this.name = 'BodyRequiredError';
+    }
+}