+++ apps/analysis-api/src/errors/invalid-json.error.ts
@@ -1,0 +1,10 @@
+import { AbstractError } from './abstract.error';
+
+export class InvalidJsonError extends AbstractError {
+
+    constructor() {
+        super('Invalid JSON string to parse', 400);
+        this.name = 'InvalidJsonError';
+    }
+
+}