+++ apps/property-management-api/src/app.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { HttpModule } from '@nestjs/axios';
 import { loadConfig } from './config';
 // TODO: import { RequestsModule } from './requests/requests.module';
 
@@ -9,7 +10,8 @@ import { loadConfig } from './config';
             isGlobal: true,
             load: [loadConfig],
         }),
+        HttpModule,
         // TODO: RequestsModule
-    ],
+    ]
 })
 export class AppModule {}