+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-05-31
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Keyword config for all words or tokens to define priority level.
+- [Marek Kurańda](https://github.com/mjkuranda): `AnalysisRequestDTO` object with validation and transformation.
+- [Marek Kurańda](https://github.com/mjkuranda): Custom error different classes.
+- [Marek Kurańda](https://github.com/mjkuranda): `AnalysisService` to analyse request and `ValidationService` to validate request body.
+
 ## [1.0.0] - 2025-05-30
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Initial basic API version.